test(cli): harden mutable config cleanup coverage
