+++ test/repro-2681-group-writable.test.ts
@@ -12,13 +12,20 @@
  * shields are up (root-owned), startup must not weaken the lock.
  */
 
-import { spawnSync } from "node:child_process";
+import { type SpawnSyncOptionsWithStringEncoding, spawnSync } from "node:child_process";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 import { describe, expect, it, vi } from "vitest";
 
 const START_SCRIPT = path.join(import.meta.dirname, "..", "scripts", "nemoclaw-start.sh");
+const MUTABLE_CONFIG_NORMALIZER = path.join(
+  import.meta.dirname,
+  "..",
+  "scripts",
+  "lib",
+  "normalize_mutable_config_perms.py",
+);
 const OPENCLAW_CONFIG_GUARD = "/usr/local/lib/nemoclaw/openclaw-config-guard.py";
 const STATE_DIR_GUARD = "/usr/local/lib/nemoclaw/state-dir-guard.py";
 const HERMES_RUNTIME_CONFIG_GUARD = "/usr/local/lib/nemoclaw/hermes-runtime-config-guard.py";
@@ -51,6 +58,43 @@ function normalizeMutableConfigPermsFor(configDir: string): string {
 
 function modeBits(filePath: string): number {
   return fs.statSync(filePath).mode;
+}
+
+function runMutableConfigNormalizer(configDir: string, ownedPaths: string[]) {
+  const testRoot = path.dirname(configDir);
+  const normalizerPath = path.join(testRoot, "normalize_mutable_config_perms.py");
+  fs.copyFileSync(MUTABLE_CONFIG_NORMALIZER, normalizerPath);
+  fs.chmodSync(normalizerPath, 0o755);
+  const spawnOptions: SpawnSyncOptionsWithStringEncoding = {
+    encoding: "utf-8",
+    env: {
+      ...process.env,
+      BASH_ENV: "",
+      HOME: testRoot,
+      NEMOCLAW_MUTABLE_CONFIG_NORMALIZER: normalizerPath,
+    },
+    timeout: 5000,
+  };
+  if (process.getuid?.() === 0) {
+    const unprivilegedId = 65534;
+    for (const ownedPath of [...ownedPaths, normalizerPath]) {
+      fs.chownSync(ownedPath, unprivilegedId, unprivilegedId);
+    }
+    spawnOptions.uid = unprivilegedId;
+    spawnOptions.gid = unprivilegedId;
+  }
+  return spawnSync(
+    "bash",
+    [
+      "-c",
+      [
+        "set -euo pipefail",
+        normalizeMutableConfigPermsFor(configDir),
+        "normalize_mutable_config_perms",
+      ].join("\n"),
+    ],
+    spawnOptions,
+  );
 }
 
 function withMockedDockerExecFileSync<T>(
@@ -247,22 +291,15 @@ describe("mutable agent config permissions", () => {
       fs.chmodSync(nestedDir, 0o700);
       fs.chmodSync(configFile, 0o600);
 
-      const result = spawnSync(
-        "bash",
-        [
-          "-c",
-          [
-            "set -euo pipefail",
-            'id() { if [ "${1:-}" = "-u" ]; then printf "1000"; else command id "$@"; fi; }',
-            'stat() { if [ "${1:-}" = "-c" ] && [ "${2:-}" = "%U" ]; then printf "sandbox\\n"; else command stat "$@"; fi; }',
-            normalizeMutableConfigPermsFor(configDir),
-            "normalize_mutable_config_perms",
-          ].join("\n"),
-        ],
-        { encoding: "utf-8", timeout: 5000 },
-      );
+      const result = runMutableConfigNormalizer(configDir, [
+        tmpDir,
+        configDir,
+        path.join(configDir, "agents"),
+        nestedDir,
+        configFile,
+      ]);
 
-      expect(result.status).toBe(0);
+      expect(result.status, result.stderr).toBe(0);
       expect(modeBits(configDir) & 0o7777).toBe(0o2770);
       expect(modeBits(configFile) & 0o7777).toBe(0o660);
       expect(modeBits(configDir) & 0o070).toBe(0o070);
@@ -302,22 +339,16 @@ describe("mutable agent config permissions", () => {
       expect(modeBits(configDir) & 0o7777).toBe(0o700);
       expect(modeBits(configFile) & 0o7777).toBe(0o600);
 
-      const result = spawnSync(
-        "bash",
-        [
-          "-c",
-          [
-            "set -euo pipefail",
-            'id() { if [ "${1:-}" = "-u" ]; then printf "1000"; else command id "$@"; fi; }',
-            'stat() { if [ "${1:-}" = "-c" ] && [ "${2:-}" = "%U" ]; then printf "sandbox\\n"; else command stat "$@"; fi; }',
-            normalizeMutableConfigPermsFor(configDir),
-            "normalize_mutable_config_perms",
-          ].join("\n"),
-        ],
-        { encoding: "utf-8", timeout: 5000 },
-      );
+      const result = runMutableConfigNormalizer(configDir, [
+        tmpDir,
+        configDir,
+        path.join(configDir, "agents"),
+        nestedDir,
+        configFile,
+        hashFile,
+      ]);
 
-      expect(result.status).toBe(0);
+      expect(result.status, result.stderr).toBe(0);
       // Mutable contract restored: setgid + group rwx dir, group rw files.
       expect(modeBits(configDir) & 0o7777).toBe(0o2770);
       expect(modeBits(configFile) & 0o7777).toBe(0o660);
